Created auth.php and tested OAuth workflow
Wrote OAuth workflow and checking script for command line in
auth.php. This code will register, validate and test tweet from the
command line (not checked in the browser, but it doesn't output any
html anyway). keys and secrets are _not_ securely stored, they are
just output to files.

This code can now be massaged into the manager script to start the
OAuth implementation.